requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
six==1.15.0
SQLAlchemy==1.4.11
Werkzeug==1.0.1
